+++ net-reporter/src/query/charts/http_request_methods_distribution/request/requester.rs
@@ -1,5 +1,8 @@
 use std::sync::Arc;
 
+use net_reporter_api::api::http_request_methods_distribution::http_request_methods_distribution::HttpRequestMethodsDistributionDTO;
+use net_reporter_api::api::http_request_methods_distribution::http_request_methods_distribution_request::HttpRequestMethodsDistributionRequestDTO;
+use net_reporter_api::api::http_request_methods_distribution::http_request_methods_distribution_filters::HttpRequestMethodsDisributionFiltersDTO;
 use net_token_verifier::fusion_auth::jwt_token::Jwt;
 use sqlx::types::chrono::DateTime;
 use sqlx::types::chrono::TimeZone;
@@ -8,17 +11,14 @@ use sqlx::Error;
 use sqlx::Pool;
 use sqlx::Postgres;
 
-use net_core_api::envelope::envelope::Envelope;
-use net_core_api::typed_api::Typed;
 use net_core_api::decoder_api::Decoder;
 use net_core_api::encoder_api::Encoder;
+use net_core_api::envelope::envelope::Envelope;
+use net_core_api::typed_api::Typed;
 
-use net_reporter_api::api::network_bandwidth_per_protocol::network_bandwidth_per_protocol::NetworkBandwidthPerProtocolDTO;
-use net_reporter_api::api::network_bandwidth_per_protocol::network_bandwidth_per_protocol_request::NetworkBandwidthPerProtocolRequestDTO;
-use net_reporter_api::api::network_bandwidth_per_protocol::network_bandwidth_per_protocol_filters::NetworkBandwidthPerProtocolFiltersDTO;
 
-use crate::query::charts::network_bandwidth_per_protocol::response::network_bandwidth_per_protocol::NetworkBandwidthPerProtocolResponse;
-use crate::query::charts::network_bandwidth_per_protocol::response::protocol::ProtocolResponse;
+use crate::query::charts::http_request_methods_distribution::response::http_request::HttpRequestMethodResponse;
+use crate::query::charts::http_request_methods_distribution::response::http_request_methods_distribution::HttpRequestMethodsDistributionResponse;
 use crate::query::requester::Requester;
 use crate::query_builder::query_builder::QueryBuilder;
 use crate::query_builder::sqlx_query_builder_wrapper::SqlxQueryBuilderWrapper;
@@ -39,18 +39,16 @@ const SET_UPPER_BYTES_BOUND: &str = "
     AND SUM(packet_length) < {}
 ";
 
-const NETWORK_BANDWIDTH_PER_PROTOCOL_REQUEST_QUERY: &str = "
-    SELECT SUM(packet_length) AS total_bytes, separated_protocols AS protocol_name
-    FROM (
-        SELECT *, UNNEST(STRING_TO_ARRAY(protocols, ':')) AS separated_protocols
-        FROM bandwidth_per_protocol_aggregate
-    ) AS unnested_protocols
+const HTTP_REQUEST_METHODS_QUERY: &str = "
+    SELECT http->>'http.request.method' m, COUNT(src_addr)
+    FROM http_request_methods_distribution_aggregate, jsonb_path_query(http_part, '$.*') AS http
     WHERE
         group_id = $1
         AND bucket >= $2
         AND bucket < $3
+        AND http->'http.request.method' IS NOT NULL
         {}
-    GROUP BY separated_protocols
+    GROUP BY m
     HAVING
         1 = 1
         {}
@@ -58,9 +56,9 @@ const NETWORK_BANDWIDTH_PER_PROTOCOL_REQUEST_QUERY: &str = "
 ";
 
 #[derive(Default)]
-pub struct NetworkBandwidthPerProtocolRequester {}
+pub struct HttpRequestMethodsDistributionRequester {}
 
-impl NetworkBandwidthPerProtocolRequester {
+impl HttpRequestMethodsDistributionRequester {
     pub fn boxed(self) -> Box<Self> {
         Box::new(self)
     }
@@ -71,9 +69,9 @@ impl NetworkBandwidthPerProtocolRequester {
         group_id: Option<&str>,
         start_date: DateTime<Utc>,
         end_date: DateTime<Utc>,
-        filters: &NetworkBandwidthPerProtocolFiltersDTO,
-    ) -> Result<Vec<ProtocolResponse>, Error> {
-        SqlxQueryBuilderWrapper::<ProtocolResponse>::new(query_string)
+        filters: &HttpRequestMethodsDisributionFiltersDTO,
+    ) -> Result<Vec<HttpRequestMethodResponse>, Error> {
+        SqlxQueryBuilderWrapper::<HttpRequestMethodResponse>::new(query_string)
             .add_option_param(group_id.map(|group_id| group_id.to_string()))
             .add_param(start_date)
             .add_param(end_date)
@@ -85,7 +83,7 @@ impl NetworkBandwidthPerProtocolRequester {
 }
 
 #[async_trait::async_trait]
-impl Requester for NetworkBandwidthPerProtocolRequester {
+impl Requester for HttpRequestMethodsDistributionRequester {
     async fn request(
         &self,
         connection_pool: Arc<Pool<Postgres>>,
@@ -97,15 +95,15 @@ impl Requester for NetworkBandwidthPerProtocolRequester {
         if enveloped_request.get_type() != self.get_requesting_type() {
             return Err(format!("wrong request is being received: {}", enveloped_request.get_type()));
         }
-        let request = NetworkBandwidthPerProtocolRequestDTO::decode(enveloped_request.get_data());
+        let request = HttpRequestMethodsDistributionRequestDTO::decode(enveloped_request.get_data());
         let request_start_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_start_date_time()).unwrap();
         let request_end_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_end_date_time()).unwrap();
-        let request_filters = request.get_filters();
+        let filters = request.get_filters();
 
-        let query = QueryBuilder::new(NETWORK_BANDWIDTH_PER_PROTOCOL_REQUEST_QUERY, 4)
-            .add_dynamic_filter(request_filters.is_include_endpoints_mode(), 1, INCLUDE_ENDPOINT_FILTER_QUERY, EXCLUDE_ENDPOINT_FILTER_QUERY)
-            .add_static_filter(request_filters.get_bytes_lower_bound(), SET_LOWER_BYTES_BOUND, 1)
-            .add_static_filter(request_filters.get_bytes_upper_bound(), SET_UPPER_BYTES_BOUND, 1)
+        let query = QueryBuilder::new(HTTP_REQUEST_METHODS_QUERY, 4)
+            .add_dynamic_filter(filters.is_include_endpoints_mode(), 1, INCLUDE_ENDPOINT_FILTER_QUERY, EXCLUDE_ENDPOINT_FILTER_QUERY)
+            .add_static_filter(filters.get_bytes_lower_bound(), SET_LOWER_BYTES_BOUND, 1)
+            .add_static_filter(filters.get_bytes_upper_bound(), SET_UPPER_BYTES_BOUND, 1)
             .build_query();
 
         let executed_query_response = Self::execute_query(
@@ -114,7 +112,7 @@ impl Requester for NetworkBandwidthPerProtocolRequester {
             Some(jwt.get_tenant_id()),
             request_start_date,
             request_end_date,
-            request_filters,
+            filters,
         ).await;
 
         if let Err(e) = executed_query_response {
@@ -122,20 +120,20 @@ impl Requester for NetworkBandwidthPerProtocolRequester {
         }
         let executed_query_response = executed_query_response.unwrap();
 
-        let response: NetworkBandwidthPerProtocolResponse = executed_query_response.into();
+        let response: HttpRequestMethodsDistributionResponse = executed_query_response.into();
         log::info!("Got response on request: {:?}", response);
 
-        let dto_response: NetworkBandwidthPerProtocolDTO = response.into();
+        let dto_response: HttpRequestMethodsDistributionDTO = response.into();
 
         Ok(Envelope::new(
             enveloped_request.get_jwt_token().ok(),
             request_agent_id,
-            NetworkBandwidthPerProtocolDTO::get_data_type(),
+            HttpRequestMethodsDistributionDTO::get_data_type(),
             &dto_response.encode()
         ))
     }
     
     fn get_requesting_type(&self) -> &'static str {
-        NetworkBandwidthPerProtocolRequestDTO::get_data_type()
+        HttpRequestMethodsDistributionRequestDTO::get_data_type()
     }
 }